Moved helper functions to separate php file for future use
